requirements.txt
flask
bertopic[all]
pandas
scikit-learn
umap-learn
hdbscan
